Fix some inflexibility with config/ci mode
The config options were effectively being ignored and didn't really
allow for fixtures. I added a static route. Unfortunately, because of
the architecture of the routing, you're more or less locked into setting
up your project the way the original author wanted (lib files in
public/javascript, etc).

CI mode was not returning an exit status reflective of the build status,
contrary to what most CI servers expect. By way of gross node process
hooks, I was able to facilitate that so it could be used in a CI server
like Hudson.